findbugs:gui doesn't load plugins #37
GUI gets configured plugins
